chat

+++ src/main/java/net/slipp/web/QuestionController.java
@@ -4,6 +4,8 @@ import net.slipp.domain.Question;
 import net.slipp.domain.QuestionRepository;
 import net.slipp.domain.Users;
 import org.h2.engine.Mode;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Controller;
@@ -17,6 +19,7 @@ import java.util.Optional;
 @RequestMapping("/questions")
 public class QuestionController {
 
+    private Logger logger = LoggerFactory.getLogger(QuestionController.class);
     @Autowired
     private QuestionRepository questionRepository;
 
@@ -31,7 +34,7 @@ public class QuestionController {
     @PostMapping
     public String create(String title, String contents,HttpSession session) {
         if(!HttpSessionUtils.isLoginUser(session)) {
-            return "/users//loginForm";
+            return "/users/loginForm";
         }
 
         Users sessionUser = HttpSessionUtils.getUserFromSession(session);
@@ -42,26 +45,54 @@ public class QuestionController {
 
     @GetMapping("/{id}")
     public String show(@PathVariable Long id, Model model) {
-        model.addAttribute("question",questionRepository.findById(id).orElseThrow(RuntimeException::new));
+        Question question = questionRepository.findById(id).orElseThrow(RuntimeException::new);
+        model.addAttribute("question", question);
         return "/qna/show";
     }
 
     @GetMapping("/{id}/form")
-    public String updateForm(@PathVariable Long id , Model model) {
-        model.addAttribute("question",questionRepository.findById(id).orElseThrow(RuntimeException::new));
+    public String updateForm(@PathVariable Long id , Model model , HttpSession session) {
+        if(!HttpSessionUtils.isLoginUser(session)) {
+            return "/users/loginForm";
+        }
+        Users loginUser = HttpSessionUtils.getUserFromSession(session);
+        Question question = questionRepository.findById(id).get();
+        logger.info("question {} ",question);
+        if(!question.isSameWriter(loginUser)) {
+            return "/users/loginForm";
+        }
+        model.addAttribute("question", question);
         return "/qna/updateForm";
     }
 
     @PutMapping("/{id}")
-    public String update(@PathVariable Long id, String title , String contents) {
+    public String update(@PathVariable Long id, String title , String contents, HttpSession session) {
+        if(!HttpSessionUtils.isLoginUser(session)) {
+            return "/users/loginForm";
+        }
+
         Question question = questionRepository.findById(id).orElseThrow(RuntimeException::new);
+        Users loginUser = HttpSessionUtils.getUserFromSession(session);
+        if(!question.isSameWriter(loginUser)) {
+            return "/users/loginForm";
+        }
         question.update(title,contents);
         questionRepository.save(question);
         return String.format("redirect:/questions/%d",id);
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable Long id){
+    public String delete(@PathVariable Long id,HttpSession session){
+        if(!HttpSessionUtils.isLoginUser(session)) {
+            return "/users/loginForm";
+        }
+
+        Question question = questionRepository.findById(id).orElseThrow(RuntimeException::new);
+        Users loginUser = HttpSessionUtils.getUserFromSession(session);
+
+        if(!question.isSameWriter(loginUser)) {
+            return "/users/loginForm";
+        }
         questionRepository.deleteById(id);
         return "redirect:/";
     }
